First functions
For now, the URL gets validated using a RegExp. If the URL is 'valid', it gets pinged using XMLHttpRequest and shows the status and response. Unfortunately, this doesn't work for cross-site checking. Next task is to check the responseText and analyze if the URL shows a file or a folder.