Function for local check
The required functions should now work for local files and folders. They will not work for other servers because of CORS policy. The solution would be to use a 'proxy' server to which the request gets send. The server could than ping the requested URL and give back the response asynchronously.